preload the config module
