+++ lib/humidifier/version.rb
@@ -1,3 +1,3 @@
 module Humidifier
-  VERSION = '0.0.35'.freeze
+  VERSION = '0.0.36'.freeze
 end
